rename PatternLib to UiLibrary
