+++ .storybook/preview.js
@@ -1,5 +1,5 @@
 
-import '!style-loader!css-loader!sass-loader!./../projects/pattern-lib/src/styles/global.scss';
+import '!style-loader!css-loader!sass-loader!./../projects/ui-library/src/styles/global.scss';
 import { setCompodocJson } from "@storybook/addon-docs/angular";
 import docJson from "../documentation.json";
 setCompodocJson(docJson);
