+++ README.md
@@ -1,4 +1,4 @@
-# DesignSystem
+# Radio Rasclat UI Library
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 10.1.4.
