+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "design-system",
+  "name": "radio-rasclat-ui-library",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -8,7 +8,7 @@
     "build:prod": "ng build --prod && npm run copy-assets",
     "test": "ng test",
     "lint": "ng lint",
-    "copy-assets": "cp -R ./projects/pattern-lib/src/styles dist/pattern-lib/",
+    "copy-assets": "cp -R ./projects/ui-library/src/styles dist/ui-library/",
     "docs:json": "compodoc -p ./tsconfig.json -e json -d .",
     "storybook": "npm run docs:json && start-storybook -p 6006",
     "build-storybook": "npm run docs:json && build-storybook"
