--- projects/pattern-lib/README.md
@@ -1,24 +1,0 @@
-# PatternLib
-
-This library was generated with [Angular CLI](https://github.com/angular/angular-cli) version 10.1.5.
-
-## Code scaffolding
-
-Run `ng generate component component-name --project pattern-lib` to generate a new component. You can also use `ng generate directive|pipe|service|class|guard|interface|enum|module --project pattern-lib`.
-> Note: Don't forget to add `--project pattern-lib` or else it will be added to the default project in your `angular.json` file. 
-
-## Build
-
-Run `ng build pattern-lib` to build the project. The build artifacts will be stored in the `dist/` directory.
-
-## Publishing
-
-After building your library with `ng build pattern-lib`, go to the dist folder `cd dist/pattern-lib` and run `npm publish`.
-
-## Running unit tests
-
-Run `ng test pattern-lib` to execute the unit tests via [Karma](https://karma-runner.github.io).
-
-## Further help
-
-To get more help on the Angular CLI use `ng help` or go check out the [Angular CLI README](https://github.com/angular/angular-cli/blob/master/README.md).
