--- projects/pattern-lib/src/public-api.ts
@@ -1,10 +1,0 @@
-/*
- * Public API Surface of pattern-lib
- */
-
-export * from './lib/pattern-lib.service';
-export * from './lib/pattern-lib.component';
-export * from './lib/pattern-lib.module';
-
-export * from './lib/button/button.component';
-export * from './lib/buttonset/buttonset.component';
