+++ projects/ui-library/README.md
@@ -1,0 +1,24 @@
+# Radio Rasclat UI Library
+
+This library was generated with [Angular CLI](https://github.com/angular/angular-cli) version 10.1.5.
+
+## Code scaffolding
+
+Run `ng generate component component-name --project ui-library` to generate a new component. You can also use `ng generate directive|pipe|service|class|guard|interface|enum|module --project ui-library`.
+> Note: Don't forget to add `--project ui-library` or else it will be added to the default project in your `angular.json` file. 
+
+## Build
+
+Run `ng build ui-library` to build the project. The build artifacts will be stored in the `dist/` directory.
+
+## Publishing
+
+After building your library with `ng build ui-library`, go to the dist folder `cd dist/ui-library` and run `npm publish`.
+
+## Running unit tests
+
+Run `ng test ui-library` to execute the unit tests via [Karma](https://karma-runner.github.io).
+
+## Further help
+
+To get more help on the Angular CLI use `ng help` or go check out the [Angular CLI README](https://github.com/angular/angular-cli/blob/master/README.md).
