+++ projects/ui-library/ng-package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/pattern-lib",
+  "dest": "../../dist/ui-library",
   "lib": {
     "entryFile": "src/public-api.ts",
     "styleIncludePaths": [
