+++ projects/ui-library/src/lib/ui-library.module.ts
@@ -1,15 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { PatternLibComponent } from './pattern-lib.component';
+import { UiLibraryComponent } from './ui-library.component';
 import { ButtonComponent } from './button/button.component';
 import { ButtonsetComponent } from './buttonset/buttonset.component';
 
 
 @NgModule({
-  declarations: [PatternLibComponent, ButtonComponent, ButtonsetComponent],
+  declarations: [UiLibraryComponent, ButtonComponent, ButtonsetComponent],
   imports: [
     CommonModule
   ],
-  exports: [PatternLibComponent, ButtonComponent, ButtonsetComponent]
+  exports: [UiLibraryComponent, ButtonComponent, ButtonsetComponent]
 })
-export class PatternLibModule { }
+export class UiLibraryModule { }
