+++ projects/ui-library/src/lib/ui-library.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 @Injectable({
   providedIn: 'root'
 })
-export class PatternLibService {
+export class UiLibraryService {
 
   constructor() { }
 }
