+++ projects/ui-library/src/public-api.ts
@@ -1,0 +1,10 @@
+/*
+ * Public API Surface of ui-library
+ */
+
+export * from './lib/ui-library.service';
+export * from './lib/ui-library.component';
+export * from './lib/ui-library.module';
+
+export * from './lib/button/button.component';
+export * from './lib/buttonset/buttonset.component';
