+++ stories/Button.stories.ts
@@ -1,12 +1,12 @@
 // also exported from '@storybook/angular' if you can deal with breaking changes in 6.1
 import { Story, Meta } from '@storybook/angular/types-6-0';
-import { ButtonComponent } from '../projects/pattern-lib/src/lib/button/button.component';
+import { ButtonComponent } from '../projects/ui-library/src/lib/button/button.component';
 
 export default {
   title: 'Atomics/Buttons',
   component: ButtonComponent,
   argTypes: {
-    label: { control: 'text' }, 
+    label: { control: 'text' },
     // we need to override here since in Angular it could be null as well and therefore it would become an ambigious data type for storybook
   }
 } as Meta;
