+++ stories/Buttonset.stories.ts
@@ -2,14 +2,14 @@
 import { Story, Meta } from '@storybook/angular/types-6-0';
 import { moduleMetadata } from '@storybook/angular';
 import { CommonModule } from '@angular/common';
-import { ButtonComponent } from '../projects/pattern-lib/src/lib/button/button.component';
-import { ButtonsetComponent } from '../projects/pattern-lib/src/lib/buttonset/buttonset.component';
+import { ButtonComponent } from '../projects/ui-library/src/lib/button/button.component';
+import { ButtonsetComponent } from '../projects/ui-library/src/lib/buttonset/buttonset.component';
 
 export default {
   title: 'Molecules/Button Sets',
   component: ButtonsetComponent,
   argTypes: {
-    label: { control: 'text' }, 
+    label: { control: 'text' },
     // we need to override here since in Angular it could be null as well and therefore it would become an ambigious data type for storybook
   },
   decorators: [
@@ -28,5 +28,5 @@ const Template: Story<ButtonsetComponent> = (args: ButtonsetComponent) => ({
 
 export const ButtonsetSample = Template.bind({});
 ButtonsetSample.args = {
-  
+
 };
