+++ stories/Intro.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta } from '@storybook/addon-docs/blocks';
 
-<Meta title="Library Example/Intro" />
+<Meta title="Radio Rasclat/Intro" />
 
 <style>{`
   .box {
@@ -11,7 +11,7 @@ import { Meta } from '@storybook/addon-docs/blocks';
   }
 `}</style>
 
-# Hello you!
+# Radio Rasclat UI Library
 
 <div class="box">
   This is the Pattern Library Intro File to provide a guiding path or whatever else you want to do here. Feel free!
